Add rate limiter to cache engine
This adds a simple rate limiter that tracks the last thirty (30) queries
made by the library. If the 30th previous query was made within the last
ten (10) seconds, the rate limiter will wait until that time has lapsed.
This mechanism is currently blocking, and there is no switch to make it
non-blocking.

Fixes #49